requirements.txt
pandas
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
defusedxml==0.8.0rc2
Django==5.0.2
django-cors-headers==4.3.1
django-cuser==2017.3.16
djangorestframework==3.14.0
pillow==10.2.0
requests==2.31.0
whitenoise==6.6.0
scikit-learn==1.2.2